requirements.txt
streamlit
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-generativeai
gspread
num2words
watchdog
